Add dummy line numbers
