+++ client/src/app/site/pages/meetings/pages/motions/services/common/motion-line-numbering.service/motion-line-numbering.service.ts
@@ -14,6 +14,7 @@ import {
 } from '../../../modules/change-recommendations/services';
 import { ViewMotion } from '../../../view-models';
 import { ViewMotionAmendedParagraph } from '../../../view-models/view-motion-amended-paragraph';
+import { LineNumbering } from '@openslides/motion-diff';
 
 /**
  * Describes the single paragraphs from the base motion.
@@ -408,7 +409,22 @@ export class MotionLineNumberingService {
             if (withDiff) {
                 return diff;
             } else {
-                return this.diffService.diffHtmlToFinalText(diff);
+                let changedText = this.diffService.diffHtmlToFinalText(diff);
+                const newRange = this.lineNumberingService.getLineNumberRange(changedText);
+                const origRange = this.lineNumberingService.getLineNumberRange(origText);
+                if (newRange.to < origRange.to) {
+                    let lineNumbers = ``;
+                    for (let n = Math.max(newRange.to, origRange.from); n <= origRange.to; n++) {
+                        lineNumbers += LineNumbering.getLineNumberElement(n).outerHTML;
+                    }
+                    if (changedText === ``) {
+                        changedText = `<p>${lineNumbers}</p>`;
+                    } else {
+                        changedText = `${changedText}${lineNumbers}`;
+                    }
+                }
+
+                return changedText;
             }
         });
     }
